Fix error output on login and signup
