+++ lambda_mud/package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "axios": "^0.18.0",
     "express": "^4.16.3",
-    "pusher": "^2.1.3",
+    "pusher-js": "^4.3.1",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "react-router": "^4.3.1",
