+++ lambda_mud/src/App.js
@@ -2,10 +2,10 @@ import React, { Component } from 'react'
 import { Route } from 'react-router-dom'
 import { withRouter } from 'react-router'
 import axios from 'axios'
-import Login from './components/Login'
-import Signup from './components/Signup'
-import Game from './components/Game'
-import Navbar from './components/Navbar'
+import Login from './components/authentication/Login'
+import Signup from './components/authentication/Signup'
+import Game from './components/gameplay/Game'
+import Navbar from './components/navbar/Navbar'
 
 class App extends Component {
   constructor(props) {
@@ -13,24 +13,14 @@ class App extends Component {
     this.state = {
       logged_in: sessionStorage.getItem('token') ? true : false,
       username: '',
-    }
-  }
-
-  componentDidMount() {
-    if(this.state.logged_in) {
-      axios({
-        method: 'post',
-      })
+      error: '',
     }
   }
 
   // Handles both Signup and Login by accepting appropriate URLS and arguments for each
   handleSignin = (userInfo, URL) => {
-    axios({
-      method: 'post',
-      url: URL,
-      data: userInfo,
-    }).then(res => {
+    let data = userInfo
+    axios.post( URL, data ).then(res => {
       const token = res.data.key
       sessionStorage.setItem('token', token)
       this.setState({'username': userInfo.username})
@@ -38,7 +28,8 @@ class App extends Component {
       console.log("Response", res)
     })
     .catch(err => {
-      console.log("Error", err.response.data.error)
+      this.setState({'error': err.response.data.error})
+      console.log("Error123", this.state.error)
     })
   }
 
@@ -50,9 +41,9 @@ class App extends Component {
   render() {
     return (
       <div>
-        <Route path='/' render={props => <Navbar {...props} handleLogout={this.handleLogout} username={this.state.username}/>} />
-        <Route exact path='/signup' render={props => <Signup {...props} handleSignin={this.handleSignin}/>} />
-        <Route exact path='/login' render={props => <Login {...props} handleSignin={this.handleSignin}/>} />
+        <Route path='/' render={props => <Navbar {...props} handleLogout={this.handleLogout} username={this.state.username} />} />
+        <Route exact path='/' render={props => <Signup {...props} handleSignin={this.handleSignin} error={this.state.error} />} />
+        <Route exact path='/login' render={props => <Login handleSignin={this.handleSignin} error={this.state.error} />} />
         <Route path='/game' render={props => <Game {...props} logged_in={this.state.logged_in} />} />
       </div>
     );
