+++ lambda_mud/src/components/gameplay/textOutput/LocationInfo.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+
+
+const LocationInfo = (props) => {
+    return (
+        <div>
+            Some Location info
+        </div>
+    )
+}
