+++ lambda_mud/src/components/gameplay/textOutput/Notifications.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react'
+import Styled from 'styled-components'
+
+const Text = styled.div`
+    color: green;
+    
+`
+
+const Notifications = (props) => {
+    return (
+        <div>
+            Some Notification
+        </div>
+    )
+}
